requirements.txt
watchdog>=3.0.0
python-frontmatter>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp>=0.9.0
tenacity>=8.0.0
requests>=2.31.0
